ajute geral de usuario e clinica

+++ src/use-cases/clinics/delete-clinic/delete-clinics.use-case.ts
@@ -4,8 +4,9 @@ import { ClinicsRepositoryInterface } from '@/repositories/clinics.interface'
 export class DeleteClinicsUseCase implements DeleteUseCaseInterface {
   constructor(protected readonly repository: ClinicsRepositoryInterface) {}
 
-  async execute(id: string): Promise<void> {
-    await this.repository.delete(id)
+  async execute(args: any): Promise<void> {
+    const { id, userId } = args
+    await this.repository.delete({ id, userId })
     return
   }
 }
